requirements.txt
mysql-connector-python
pandas
youtube_transcript_api
pymongo
